Refactor: Merge AuxController into main header
Merged AuxController.h into xDuinoRails_DccLightsAndFunctions.h and renamed the .cpp file to xDuinoRails_DccLightsAndFunctions.cpp.

This change simplifies the library's file structure by consolidating the main controller class into the primary library header file. No functional changes were made.
